# Use OnceLock instead of OnceCell