@mtreinish 2024-11-11
After 8bc48442 merged fixing the asv tests I did a quick asv run which yielded basically no change:

```
Benchmarks that have stayed the same:

| Change   | Before [8bc48442] <once-lock^2>   | After [487e1bec] <once-lock>   | Ratio   | Benchmark (Parameter)                                                                                           |
|----------|-----------------------------------|--------------------------------|---------|-----------------------------------------------------------------------------------------------------------------|
|          | 113±5μs                           | 246±70μs                       | ~2.17   | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 8192)                                        |
|          | 452±200μs                         | 827±300μs                      | ~1.83   | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 32768)                                       |
|          | 486±200μs                         | 799±200μs                      | ~1.64   | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 32768)                                       |
|          | 665±300μs                         | 962±200μs                      | ~1.45   | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 32768)                                       |
|          | 461±200μs                         | 614±200μs                      | ~1.33   | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 32768)                                       |
|          | 34.2±8μs                          | 45.2±20μs                      | ~1.32   | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 2048)                                        |
|          | 194±50μs                          | 248±50μs                       | ~1.28   | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 8192)                                        |
|          | 5.89±0.2ms                        | 6.88±0.4ms                     | ~1.17   | manipulate.TestCircuitManipulate.time_DTC100_twirling                                                           |
|          | 3.42±0.03s                        | 3.97±0.2s                      | ~1.16   | utility_scale.UtilityScaleBenchmarks.time_circSU2('cz')                                                         |
|          | 222±2μs                           | 252±70μs                       | ~1.14   | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 8192)                                        |
|          | 2.26±0.5ms                        | 2.50±0.6ms                     | ~1.11   | circuit_construction.CircuitConstructionBench.time_circuit_copy(20, 131072)                                     |
|          | 172±60μs                          | 189±60μs                       | ~1.10   | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 8192)                                        |
|          | 2.28±0.4ms                        | 2.51±0.5ms                     | 1.10    | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 131072)                                     |
|          | 132±10μs                          | 144±7μs                        | 1.09    | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 8192)                                       |
|          | 2.34±0.1ms                        | 2.55±0.07ms                    | 1.09    | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 131072)                                      |
|          | 2.34±0.2ms                        | 2.53±0.04ms                    | 1.08    | circuit_construction.CircuitConstructionBench.time_circuit_copy(2, 131072)                                      |
|          | 41.9±1μs                          | 44.9±1μs                       | 1.07    | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 2048)                                       |
|          | 782±10μs                          | 833±40μs                       | 1.07    | circuit_construction.CliffordSynthesis.time_clifford_synthesis(10)                                              |
|          | 2.47±0.05ms                       | 2.62±0.05ms                    | 1.06    | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 131072)                                      |
|          | 2.37±0.1ms                        | 2.52±0.06ms                    | 1.06    | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 131072)                                      |
|          | 1.13±0.03ms                       | 1.19±0.03ms                    | 1.06    | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 8192)                                      |
|          | 5.78±0.2ms                        | 6.12±0.2ms                     | 1.06    | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 32768)                                    |
|          | 375±10μs                          | 393±2μs                        | 1.05    | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 2048)                                     |
|          | 343±7μs                           | 361±2μs                        | 1.05    | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 2048)                                      |
|          | 1.39±0.01ms                       | 1.46±0.02ms                    | 1.05    | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8192)                                      |
|          | 108±6ms                           | 113±20ms                       | 1.05    | circuit_construction.ParamaterizedDifferentCircuit.time_DTC100_set_build(100, 150)                              |
|          | 34.4±2μs                          | 35.6±1μs                       | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_copy(1, 2048)                                        |
|          | 462±30μs                          | 481±40μs                       | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_copy(14, 32768)                                      |
|          | 4.52±0.2ms                        | 4.68±0.3ms                     | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 32768)                                     |
|          | 380±2μs                           | 397±5μs                        | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 2048)                                     |
|          | 1.46±0.03ms                       | 1.51±0.04ms                    | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 8192)                                      |
|          | 5.65±0.08ms                       | 5.86±0.1ms                     | 1.04    | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 32768)                                     |
|          | 3.27±0.08ms                       | 3.39±0.05ms                    | 1.04    | circuit_construction.ParamaterizedDifferentCircuit.time_DTC100_set_build(50, 10)                                |
|          | 29.4±0.2ms                        | 30.5±0.3ms                     | 1.04    | circuit_construction.ParamaterizedDifferentCircuit.time_QV100_build(50, 150)                                    |
|          | 296±3ms                           | 307±1ms                        | 1.04    | circuit_construction.ParameterizedCircuitBindBench.time_bind_params(20, 131072, 128)                            |
|          | 39.2±3μs                          | 40.5±7μs                       | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 2048)                                        |
|          | 37.1±0.5μs                        | 38.1±0.3μs                     | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 128)                                      |
|          | 28.9±0.3μs                        | 29.8±0.09μs                    | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 128)                                       |
|          | 5.32±0.1ms                        | 5.50±0.07ms                    | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 32768)                                     |
|          | 44.2±0.3μs                        | 45.4±0.4μs                     | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 128)                                      |
|          | 22.3±1ms                          | 23.0±1ms                       | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 131072)                                   |
|          | 22.9±0.3μs                        | 23.7±0.4μs                     | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 8)                                        |
|          | 364±7μs                           | 375±7μs                        | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 2048)                                      |
|          | 1.48±0.04ms                       | 1.53±0.05ms                    | 1.03    | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 8192)                                      |
|          | 4.78±0.04ms                       | 4.94±0.1ms                     | 1.03    | circuit_construction.CliffordSynthesis.time_clifford_synthesis(50)                                              |
|          | 8.91±0.3ms                        | 9.22±0.2ms                     | 1.03    | circuit_construction.ParamaterizedDifferentCircuit.time_DTC100_set_build(100, 10)                               |
|          | 34.6±0.9ms                        | 35.5±0.7ms                     | 1.03    | circuit_construction.ParamaterizedDifferentCircuit.time_DTC100_set_build(50, 150)                               |
|          | 604±4ms                           | 624±2ms                        | 1.03    | circuit_construction.ParameterizedCircuitConstructionBench.time_build_parameterized_circuit(20, 131072, 8)      |
|          | 9.40±0.06ms                       | 9.65±0.06ms                    | 1.03    | utility_scale.UtilityScaleBenchmarks.time_parse_qaoa_n100('ecr')                                                |
|          | 484±4ms                           | 495±3ms                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 131072)                             |
|          | 481±1ms                           | 490±7ms                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 131072)                              |
|          | 123±0.3ms                         | 125±0.6ms                      | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 32768)                              |
|          | 30.9±0.1ms                        | 31.5±0.05ms                    | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8192)                               |
|          | 30.8±0.2ms                        | 31.4±0.1ms                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(5, 8192)                                |
|          | 572±3μs                           | 585±2μs                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 128)                                 |
|          | 123±0.5ms                         | 125±1ms                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_construction(8, 32768)                               |
|          | 11.9±0.05μs                       | 12.0±0.1μs                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_copy(5, 128)                                         |
|          | 12.8±0.4μs                        | 13.1±0.2μs                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_copy(8, 128)                                         |
|          | 25.3±0.05μs                       | 25.9±0.1μs                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 128)                                       |
|          | 293±3μs                           | 297±8μs                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(1, 2048)                                      |
|          | 1.53±0.04ms                       | 1.56±0.04ms                    | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(14, 8192)                                     |
|          | 8.80±0.1μs                        | 8.93±0.06μs                    | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(2, 8)                                         |
|          | 5.94±0.1ms                        | 6.04±0.1ms                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 32768)                                    |
|          | 1.48±0.03ms                       | 1.51±0.06ms                    | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(20, 8192)                                     |
|          | 32.6±0.4μs                        | 33.4±0.2μs                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 128)                                       |
|          | 23.1±0.8ms                        | 23.6±0.6ms                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 131072)                                    |
|          | 5.65±0.09ms                       | 5.78±0.1ms                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 32768)                                     |
|          | 10.1±0.07μs                       | 10.3±0.07μs                    | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(5, 8)                                         |
|          | 22.3±0.8ms                        | 22.7±0.8ms                     | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 131072)                                    |
|          | 371±2μs                           | 380±9μs                        | 1.02    | circuit_construction.CircuitConstructionBench.time_circuit_extend(8, 2048)                                      |
|          | 12.2±0.4ms                        | 12.5±0.05ms                    | 1.02    | circuit_construction.ParamaterizedDifferentCircuit.time_DTC100_set_build(50, 50)                                |
|          | 599±3ms                           | 613±5ms                        | 1.02    | circuit_construction.ParameterizedCircuitBindBench.time_bind_params(20, 131072, 131072)                         |
|          | 145±1ms                           | 147±2ms                        | 1.02    | circuit_construction.ParameterizedCircuitBindBench.time_bind_params(20, 32768, 32768)                           |
|          | 31.4±0.5ms                        | 32.0±1ms                       | 1.02    | circuit_construction.ParameterizedCircuitBindBench.time_bind_params(20, 8192, 8192)                             |
|          | 660±4ms                           | 675±7ms                        | 1.02    | circuit_construction.ParameterizedCircuitConstructionBench.time_build_parameterized_circuit(20, 131072, 8192)   |
|          | 309±2ms                           | 314±3ms                        | 1.02    | circuit_construction.ParameterizedCircuitConstructionBench.time_build_parameterized_circuit(20, 32768, 32768)   |
|          | 38.5±0.2ms                        | 39.5±0.3ms                     | 1.02    | circuit_construction.ParameterizedCircuitConstructionBench.time_build_parameterized_circuit(20, 8192, 128)      |
|          | 46.9±0.1ms                        | 47.7±0.2ms                     | 1.02    | circuit_construction.ParameterizedCircuitConstructionBench.time_build_parameterized_circuit(20, 8192, 2048)     |
|          | 3.59±0.1s                         | 3.67±0.1s                      | 1.02    | utility_scale.UtilityScaleBenchmarks.time_circSU2('cx')                                                         |
|          | 3.60±0.08s                        | 3.65±0.2s                      | 1.02    | utility_scale.UtilityScaleBenchmarks.time_circSU2('ecr')                                                        |
|          | 100±2ms                           | 102±2ms                        | 1.02    | utility_scale.UtilityScaleBenchmarks.time_parse_qft_n100('cx')                                                  |
|          | 32.8±0.2ms                        | 33.3±0.1ms                     | 1.02    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('cx')                                    |
|          | 32.6±0.5ms                        | 33.2±0.3ms                     | 1.02    | utility_scale.UtilityScaleBenchmarks.time_parse_square_heisenberg_n100('ecr')                                   |
|          | 153±1ms                           | 156±0.9ms                      | 1.02    | utility_scale.UtilityScaleBenchmarks.time_square_heisenberg('cz')                                               |
|          | 7.29±0.01ms                       | 7.33±0.1ms                     | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(1, 2048)                                |
|          | 580±5μs                           | 587±1μs                        | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 128)                                |
|          | 158±0.5μs                         | 160±1μs                        | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8)                                  |
|          | 30.8±0.08ms                       | 31.2±0.3ms                     | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(14, 8192)                               |
|          | 30.4±0.3ms                        | 30.8±0.2ms                     | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(2, 8192)                                |
|          | 674±2μs                           | 682±2μs                        | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 128)                                |
|          | 486±3ms                           | 490±7ms                        | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 131072)                             |
|          | 7.83±0.04ms                       | 7.93±0.04ms                    | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 2048)                               |
|          | 218±0.6μs                         | 221±1μs                        | 1.01    | circuit_construction.CircuitConstructionBench.time_circuit_construction(20, 8)                                  |
…